# Christmas in August
Today's an exciting day! We'll be tackling an entirely new set of code
katas using an entirely different language! We'll be working on the
[first challenge](http://adventofcode.com/day/1) in the [Advent of
Code](http://adventofcode.com/) series.

In the same vein as one my our [recent
posts](http://www.east5th.co/blog/2016/08/15/meteor-in-front-phoenix-in-back-part-1/),
we'll be using the [Elixir](http://elixir-lang.org/) language to solve
this challenge.

This first commit is the result of `mix new ...` and creates a base
Elixir project.